// contracts/src/enum/transaction-category.enum.ts
export enum TransactionCategory {
    SALARY = 'SALARY',
    OVERTIME = 'OVERTIME',
    FREELANCE = 'FREELANCE',
    COMMISSION = 'COMMISSION',
    BONUS = 'BONUS',
    INVESTMENT = 'INVESTMENT',
    BUSINESS = 'BUSINESS',
    GIFT = 'GIFT',
    FOOD = 'FOOD',
    TRANSPORT = 'TRANSPORT',
    BILLS = 'BILLS',
    SHOPPING = 'SHOPPING',
    HEALTH = 'HEALTH',
    SAVINGS = 'SAVINGS',
    SPORT = 'SPORT',
    ENTERTAINMENT = 'ENTERTAINMENT',
    CREDIT_CARD = 'CREDIT_CARD',
    HOUSING = 'HOUSING',
    OTHER = 'OTHER',
}